Initial commit of Depths, a game written in Fathom.
As we develop Depths, we'll move the content that we feel is generally
applicable into Fathom so that we a) only add things that are actually
broadly useful and b) don't add extraneous stuff in. I will mark the
functions that I feel fit this criteria with # MOVETO FATHOM.

Right now, depths just creates a canvas and spins up a loop. Nothing too
exciting...yet. Stay tuned.